Refactor Seata's nativeTest integrated under GraalVM Native Image
