+++ test/native/src/test/java/org/apache/shardingsphere/test/natived/jdbc/commons/TestShardingService.java
@@ -81,7 +81,7 @@ public final class TestShardingService {
         assertThat(orderRepository.selectAll(), equalTo(new ArrayList<>()));
         assertThat(orderItemRepository.selectAll(), equalTo(new ArrayList<>()));
         assertThat(addressRepository.selectAll(), equalTo(new ArrayList<>()));
-        addressRepository.assertRollbackWithTransactions();
+        orderItemRepository.assertRollbackWithTransactions();
     }
     
     /**
