+++ test/native/src/test/java/org/apache/shardingsphere/test/natived/jdbc/commons/repository/OrderItemRepository.java
@@ -28,6 +28,9 @@ import java.sql.Statement;
 import java.util.LinkedList;
 import java.util.List;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+
 @SuppressWarnings({"SqlDialectInspection", "SqlNoDataSourceInspection"})
 public final class OrderItemRepository {
     
@@ -221,5 +224,31 @@ public final class OrderItemRepository {
             }
         }
         return result;
+    }
+    
+    /**
+     * Assert rollback with transactions.
+     * This is currently just a simple test against a non-existent table and does not involve the competition scenario of distributed transactions.
+     *
+     * @throws SQLException An exception that provides information on a database access error or other errors.
+     */
+    public void assertRollbackWithTransactions() throws SQLException {
+        Connection connection = dataSource.getConnection();
+        try {
+            connection.setAutoCommit(false);
+            connection.createStatement().executeUpdate("INSERT INTO t_order_item (order_id, user_id, phone, status) VALUES (2024, 2024, '13800000001', 'INSERT_TEST')");
+            connection.createStatement().executeUpdate("INSERT INTO t_order_item_does_not_exist (test_id_does_not_exist) VALUES (2024)");
+            connection.commit();
+        } catch (SQLException e) {
+            connection.rollback();
+        } finally {
+            connection.setAutoCommit(true);
+            connection.close();
+        }
+        try (
+                Connection conn = dataSource.getConnection();
+                ResultSet resultSet = conn.createStatement().executeQuery("SELECT * FROM t_order_item WHERE user_id = 2024")) {
+            assertThat(resultSet.next(), is(false));
+        }
     }
 }
